bank-backend/auth_service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
PyJWT==2.9.0
pydantic==2.8.2
